refactor: fix excessive request internals leakage
Forgery protection was reaching into Request internals to change state.
This patch moves those things into a Request#nullify_state method.

Note, this is a precursor to fixing the fact that Request may not
contain the #flash method depending on if the file is loaded.
